chore: rename project to Eurekapp
