+++ app/src/main/AndroidManifest.xml
@@ -10,18 +10,18 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.SampleApp"
+        android:theme="@style/Theme.Eurekapp"
         tools:targetApi="31">
         <activity
             android:name=".SecondActivity"
             android:exported="false"
             android:label="@string/title_activity_second"
-            android:theme="@style/Theme.SampleApp" />
+            android:theme="@style/Theme.Eurekapp" />
         <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/title_activity_main"
-            android:theme="@style/Theme.SampleApp">
+            android:theme="@style/Theme.Eurekapp">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
