+++ app/src/main/java/ch/eureka/eurekapp/resources/C.kt
@@ -1,4 +1,4 @@
-package com.android.sample.resources
+package ch.eureka.eurekapp.resources
 
 // Like R, but C
 object C {
