+++ app/src/main/java/ch/eureka/eurekapp/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package ch.eureka.eurekapp.ui.theme
 
 import androidx.compose.ui.graphics.Color
