+++ app/src/test/java/ch/eureka/eurekapp/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package ch.eureka.eurekapp
 
 import org.junit.Assert.*
 import org.junit.Test
