+++ app/src/test/java/ch/eureka/eurekapp/screen/SecondScreen.kt
@@ -1,7 +1,7 @@
-package com.android.sample.screen
+package ch.eureka.eurekapp
 
 import androidx.compose.ui.test.SemanticsNodeInteractionsProvider
-import com.android.sample.resources.C
+import ch.eureka.eurekapp.resources.C
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import io.github.kakaocup.compose.node.element.KNode
